requirements.txt
requests
python-dotenv
pandas
pyotp
schedule
customtkinter
